converted from long to double

+++ Filters/LinearKalmanFilter/LinearKalmanFilter.h
@@ -2,19 +2,19 @@
 #define LINEARKALMANFILTER_H
 
 typedef struct {
-    long* F; // State Transition Matrix
-    long* G; // Control Matrix
-    long* H; // Observation Matrix
-    long* P; // Estimate Covariance Matrix
-    long* R; // Measurement Uncertainty Matrix
-    long* K; // Kalman Gain Matrix
-    long* Q; // Process Noise Matrix
-    long* U; // Control Vector
-    long* X; // State Vector
+    double* F; // State Transition Matrix
+    double* G; // Control Matrix
+    double* H; // Observation Matrix
+    double* P; // Estimate Covariance Matrix
+    double* R; // Measurement Uncertainty Matrix
+    double* K; // Kalman Gain Matrix
+    double* Q; // Process Noise Matrix
+    double* U; // Control Vector
+    double* X; // State Vector
 
 } KFState;
 
-KFState initialize(int statevector_size, int measurement_size, int control_size, long* initial_state, long* initial_control){
+KFState initialize(int statevector_size, int measurement_size, int control_size, double* initial_state, double* initial_control){
  
 }
 void predict_state(KFState *state){
@@ -33,7 +33,7 @@ void covariance_update(KFState *state){
     
 }
 
-long* iterate(KFState *state, float dt, long* measurement_vector, long* control_vector){
+double* iterate(KFState *state, float dt, double* measurement_vector, double* control_vector){
 
 }
